+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+Bugfixes:
+- Fixed gauge connection flow animation won't play without made at least one request attempt.
+
 ## 1.0.0
 IMPORTANT: this version changed Factory Controller block data structure, once upgraded to this version, your factory controller cannot downgrade to previous mod version, so backup before upgrade.